% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLockdownDateTime.R
\name{addLockdownDateTime}
\alias{addLockdownDateTime}
\title{\code{addLockdownDateTime} takes a plot where the x axis is a daterTime and adds a 
rectangle to denote lockdown.}
\usage{
addLockdownDateTime(p, yMin, yMax)
}
\arguments{
\item{p}{the plot}

\item{yMin}{the min y value of the plot (we -1 for visibility)}

\item{yMax}{the max y value of the plot (we +1 for visibility)}
}
\description{
It assumes `myParams$lockDownStartDate` and `myParams$lockDownEndDate` define 
the start & end. Yes, they should be parameters.
}
\details{
It also assumed `myParams$myAlpha`and `myParams$vLineCol` exist.
}
\seealso{
Other plots: 
\code{\link{addLockdownDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
