% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeekendsDate.R
\name{addWeekendsDate}
\alias{addWeekendsDate}
\title{\code{addWeekendsDate} adds rectangles around the weekends from March 2020 onwards
to plots where the x-axis is a date}
\usage{
addWeekendsDate(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addWeekendsDate} adds rectangles around the weekends from March 2020 onwards
to plots where the x-axis is a date
}
\seealso{
Other plots: 
\code{\link{addLockdownRectDateTime}()},
\code{\link{addLockdownRectDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
