% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeekendsDateTime.R
\name{addWeekendsDateTime}
\alias{addWeekendsDateTime}
\title{\code{addWeekendsDateTime} adds rectangles around the weekends from March 2020 onwards
to plots where the x-axis is a dateTime.}
\usage{
addWeekendsDateTime(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addWeekendsDateTime} adds rectangles around the weekends from March 2020 onwards
to plots where the x-axis is a dateTime.
}
\seealso{
Other plots: 
\code{\link{addLockdownRectDateTime}()},
\code{\link{addLockdownRectDate}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
