% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareYearsDiffPlotDaily.R
\name{compareYearsDiffPlotDaily}
\alias{compareYearsDiffPlotDaily}
\title{\code{compareYearsDiffPlotDaily} makes a geom_step() plot from the dt passed in by
aggregating dt to dates.}
\usage{
compareYearsDiffPlotDaily(dt)
}
\arguments{
\item{dt}{the data to plot}
}
\description{
It assumes `compareYear` exists and has the values "2017-2019" and "2020". It compares the `value`
variable (doesn't care what it is) and prints min/max difference for info. 
Should use scales::percent() but doesn't.
}
\seealso{
Other plots: 
\code{\link{addLockdownDateTime}()},
\code{\link{addLockdownDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
