% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareYearsDiffPlotWeekly.R
\name{compareYearsDiffPlotWeekly}
\alias{compareYearsDiffPlotWeekly}
\title{\code{compareYearsDiffPlotWeekly} makes a geom_step() plot from the dt passed in by
aggregating dt to weeks.}
\usage{
compareYearsDiffPlotWeekly(dt, ldStart, ldEnd)
}
\arguments{
\item{dt}{the data to plot}

\item{ldStart}{lockdown start date (for annotation)}

\item{ldEnd}{lockdown end date (for annotation)}
}
\description{
It assumes `compareYear` exists and has the values "2017-2019" and "2020". It compares the `value`
variable (doesn't care what it is) and prints min/max difference for info. 
Should use scales::percent() but doesn't.
}
\seealso{
Other plots: 
\code{\link{addLockdownRectDateTime}()},
\code{\link{addLockdownRectDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
