% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareYearsPlot.R
\name{compareYearsPlot}
\alias{compareYearsPlot}
\title{\code{compareYearsPlot} makes a geom_line() plot using `colVar` to 
colour the lines. It assumes the x axis is a date.}
\usage{
compareYearsPlot(dt, xVar, yVar, colVar)
}
\arguments{
\item{dt}{the data to plot}

\item{xVar}{the variable to use for the x axis}

\item{yVar}{the variable to use for the y axis}

\item{colVar}{the variable to use to colour the tile fill}
}
\description{
\code{compareYearsPlot} makes a geom_line() plot using `colVar` to 
colour the lines. It assumes the x axis is a date.
}
\seealso{
Other plots: 
\code{\link{addLockdownRectDateTime}()},
\code{\link{addLockdownRectDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
