% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDotPlot.R
\name{makeDotPlot}
\alias{makeDotPlot}
\title{\code{makeDotPlot} makes a geom_point() x/y plot using the viridis colour scheme to 
colour byVar.}
\usage{
makeDotPlot(dt, xVar, xLab, yVar, yLab, byVar)
}
\arguments{
\item{dt}{the data to plot}

\item{xVar}{the variable to use for the x axis}

\item{xLab}{the x axis label}

\item{yVar}{the variable to use for the y axis}

\item{yLab}{the y axis label}

\item{byVar}{the variable to use to colour different points}
}
\description{
\code{makeDotPlot} makes a geom_point() x/y plot using the viridis colour scheme to 
colour byVar.
}
\seealso{
Other plots: 
\code{\link{addLockdownRectDateTime}()},
\code{\link{addLockdownRectDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeTilePlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
