% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTilePlot.R
\name{makeTilePlot}
\alias{makeTilePlot}
\title{\code{makeTilePlot} makes a geom_tile() plot using `fillVar` to 
colour the tiles.}
\usage{
makeTilePlot(dt, xVar, xLab, yVar, yLab, fillVar)
}
\arguments{
\item{dt}{the data to plot}

\item{xVar}{the variable to use for the x axis}

\item{xLab}{the x axis label}

\item{yVar}{the variable to use for the y axis}

\item{yLab}{the y axis label}

\item{fillVar}{the variable to use to colour the tile fill}
}
\description{
\code{makeTilePlot} makes a geom_tile() plot using `fillVar` to 
colour the tiles.
}
\seealso{
Other plots: 
\code{\link{addLockdownDateTime}()},
\code{\link{addLockdownDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{profileTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
