% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileTilePlot.R
\name{profileTilePlot}
\alias{profileTilePlot}
\title{\code{profileTilePlot} makes a `geom_tile()` plot from the dt passed in.}
\usage{
profileTilePlot(dt, yLab)
}
\arguments{
\item{dt}{the data.table to plot}

\item{yLab}{the y axis label}
}
\description{
It assumes `obsDate` exists and uses this for the x axis with `time` (in hms 
format) as the y. It colours the tile using the `dt$value` variable.
}
\seealso{
Other plots: 
\code{\link{addLockdownRectDateTime}()},
\code{\link{addLockdownRectDate}()},
\code{\link{addWeekendsDateTime}()},
\code{\link{addWeekendsDate}()},
\code{\link{compareYearsDiffPlotDaily}()},
\code{\link{compareYearsDiffPlotWeekly}()},
\code{\link{compareYearsPlot}()},
\code{\link{makeDotPlot}()},
\code{\link{makeTilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plots}
